requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0
sqlalchemy>=2.0.25
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0